Feature/recur (#3)
Add initial implementation of date logic.

The yaml from a trello card is interpreted as a set of recurring
dates. The "current" date is then checked against this set to
decide whether a recurring card should be created on this day.

Current implementation only supports multiples of weeks and months
as the recurring period, and only a single day within the period on
which a new card is created.

A "begin" day is added as a field in the yaml - this is required
for the logic to make sense, since e.g. day 1 of a fortnightly
cycle only makes sense given a starting date. The field is optional
and 01/01/2019 is used as the default.